refactor: alterando o banco de h2 para mysql com docker

+++ src/main/kotlin/com/apiRest/forum/dto/RespostasView.kt
@@ -3,7 +3,7 @@ package com.apiRest.forum.dto
 import java.time.LocalDateTime
 
 data class RespostasView (
-    val id: Long?,
+    val id: Int?,
     var mensagem: String,
     val nomeAutor: String,
     val dataCriacao: LocalDateTime = LocalDateTime.now(),
